backend/requirements.txt
# Google Vision API
google-cloud-vision==2.7.0

# Google Text-to-Speech API
google-cloud-texttospeech==2.5.0

# Requests
requests==2.26.0 

# Together AI API
together==0.2.6

# Firebase libraries
firebase-admin==5.3.0

# Playsound
Playsound==1.3.0
